[ Refactor ] Move manage object model helpers to extension
