+++ FeedStoreChallenge/Infrastructure/ManagedFeedImage.swift
@@ -15,6 +15,19 @@ internal class ManagedFeedImage: NSManagedObject {
 	@NSManaged var location: String?
 	@NSManaged var url: URL
 	@NSManaged var cache: ManagedCache
+}
+
+extension ManagedFeedImage {
+	static func images(feed: [LocalFeedImage], in context: NSManagedObjectContext) -> NSOrderedSet {
+		return NSOrderedSet(array: feed.map({ local in
+			let managedFeedImage = ManagedFeedImage(context: context)
+			managedFeedImage.id = local.id
+			managedFeedImage.imageDescription = local.description
+			managedFeedImage.location = local.location
+			managedFeedImage.url = local.url
+			return managedFeedImage
+		}))
+	}
 
 	var local: LocalFeedImage {
 		return LocalFeedImage(id: id, description: imageDescription, location: location, url: url)
